Create a list of Mod objects from response from server

+++ FicsitExplorer/Mod.cs
@@ -4,7 +4,7 @@
     {
             public string Name { get; set; }
             public string Description { get; set; }
-            public uint Downloads { get; set; }
+            public long Downloads { get; set; }
             public string ID { get; set; }
             public string LogoURL { get; set; } //TODO: Decide if keep as string or change to URL
     }
